Introducing the Logging Module
The new IRCLogging Module uses Python's logging module and creates the
basic configuration with a special text formatting for it. The file
structure is created automatically so that the logging module can write
to its file. IRCLogging also creates a new logfile for each day, and a
new folder for each month, even while it is running, and switches to the
correct file. It adds no new interfaces by itself, Python's logging
module has to be used to add new messages to the log file. Take a look
at the official documentation for the module, e.g.
http://docs.python.org/2/howto/logging.html#logging-basic-tutorial